src/swarm_dashboard/requirements.txt
python-dotenv==1.1.0
fastapi==0.115.12
uvicorn==0.34.2
jinja2==3.1.6
httpx==0.28.1
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-instrumentation-fastapi==0.54b1
azure-monitor-opentelemetry-exporter==1.0.0b35
